Fix password manager for forms with no username
In 8036139d, I once again broke the password manager for forms with no
username. I think I will continue to break the password manager every
time I make any change to it, so long as we don't have any tests.

This reverts 8036139d and ensures we pass along a null formAuth when
there are no results, like we used to do.