ephy-window: Execute CtrlTab shortcuts regardless of focused widget
If the focused widget was the location entry or the WebKitWebView then
CtrlTab would be handled properly, but if there was another widget focused
for example the hamburger button then the first CtrlTab would not be handled.
Instead what would happen was that the focused widget would be changed until
it would be the location entry and only then CtrlTab would be handled properly.

Fixes https://gitlab.gnome.org/GNOME/epiphany/issues/911