Fix invalid free in ephy_web_application_setup_from_desktop_file()
We are clobbering the GDesktopAppInfo's internal icon here, leading to a
critical later when it gets finalized.

Note: this function is only called when a desktop file basename is
passed to --application-mode. Epiphany will never create such web apps
itself, which is how this went unnoticed for so long, but it's used by
fedora-developer-portal (prior to F32, which switched to Chromium
instead of Epiphany at my request).