ephy-window: execute Ctrl+Tab shortcuts regardless of focused widget
If the focused widget was the location entry or the WebKitWebView then
Ctrl+Tab would be handled properly, but if there was another widget focused
for example the hamburger button then the first Ctrl+Tab would not be handled.

Instead what would happen was that the focused widget would be changed until
it would be the location entry and only then Ctrl+Tab would be handled properly.

https://gitlab.gnome.org/GNOME/epiphany/issues/911