css: Refactor Adwaita recoloring
Generalize the incognito style and move it to a separate file. Use
variables from Adwaita to not trigger special-casing in button mixing
(leading to wrong borders).

This allows to use it for both incognito and automation modes, with
different base color and icon variables.

Remove the shared incognito mode definitions other than icon and the base
color, as they are unused now.