Assert webapp is non-NULL where required
ephy_web_application_for_profile_directory() had better not return NULL
if called for the current profile directory in app mode. In this case,
we should crash as cleanly as possible, with g_assert().

But in other situatitons, it's OK for this function to return NULL.
Other callers are already prepared to handle this.