bookmarks-import: Fix importing duplicate bookmarks from HTML
Currently when importing bookmarks with the same url, but in different folders, there is a bug that causes every subsequent bookmark in the HTML file to be added with mismatching name and url.

Try to fix this by adding the latter duplicate bookmark's tag to the former duplicate, and skipping adding the bookmark's url, date and title to respective pointer arrays.