bookmarks-import: Fix importing duplicate bookmarks from HTML
Currently when importing bookmarks with the same url, but in different folders, there is a bug that causes every subsequent bookmark in the HTML file to be added with mismatching name and url.

Try to fix this by adding every duplicate bookmark with folders converted to tags. This results in duplicate entries after importing, which disappear after restart, leaving all bookmarks correctly imported with appropriate tags.