set-image-as-wallpaper: use libportal
Currently, the code that sets an image as wallpaper sets the key
in gnome-shell's gsettings, which means this only work for GNOME
and it doesn't work at all in flatpaks.

So change to use libportal instead, making this work everywhere,
even outside flatpak, because libportal works out of the sandbox
as well.

Fixes #1198