Cleanup font preference callbacks
There's a bunch of unused code here designed to prevent this function
from doing anything if called for a non-Epiphany settings schema, but
this never happens. Maybe in the past it was hooked up to watch for
changes to system settings, but if so, that was probably long, long ago.

Also, default font size should be 12.