WIP: Allow per-site autoplay policies.
Denying autoplay can break websites that use it for stylistic
effects. On the other hand, it can be increadibly annoying to be
spammed with videos autoplaying. Allow users to configure this via
GSettings.

By default, videos can autoplay so long as they have no audio (or the
audio is muted). Alternatively, users can allow all videos to autoplay
even with audio, or to deny all attempts to autoplay video.

This is WIP for two reasons,

 1. We need to wait for 2.30 to be released for the WK API to be
 available.

 2. I've messed up the combo box UI by adding this longer "Default"
 string for autoplay, instead of "Ask", so the alignment is
 buggered. Sure there's a magic number somewhere I can change to fix
 this :-).