sync-service: port to SecretPassword API
The sync service has been broken since we removed D-Bus access to the
secret service, because we forgot to port it to the new SecretPassword
API, which allows use of the secrets portal.

Hopefully fixes #1233, although I haven't tested this because the
easiest way to test flatpak-related commits is to yolo commit to master
and then check if Ephy Tech Preview still works after it is built.