search-engine-dialog: make +/- buttons linked and fill all space
Currently, the add and remove buttons are next together but have
both rounded corners everywhere which make them look weird. They
are also at small size (i.e. not taking the whole width of the
GtkListBox attached to it) and cranked on the left side.

This commit make the buttons have a linked style (no more round
corners for their common side) and take the whole width of the
GtkListBox.