prefs-search-engines: replace dialog with inline list box
The old search engines dialog had a terrible look so this commit
replaces it with a list box with expandable rows, that let see
and edit the various informations (name, address, etc.). It also
allows to select the default search engine, used when no engine
is selected when doing a search in the search bar.

Fixes #631
Fixes #441