Allow per-site autoplay policies.
Denying autoplay can break websites that use it for stylistic
effects. On the other hand, it can be increadibly annoying to be
spammed with videos autoplaying. Allow users to configure this via
GSettings.

By default, videos can autoplay so long as they have no audio (or the
audio is muted). Alternatively, users can allow all videos to autoplay
even with audio, or to deny all attempts to autoplay video.