Use getentropy(2) on OpenBSD
Linux and FreeBSD provided getrandom(2) which doesn't exist on OpenBSD,
so adjust the code for OpenBSD to the getentropy(2) instead.

While here adjust the headers to give this a chance to build on FreeBSD
too. Untested on FreeBSD but based on the manpage unistd.h provided
getrandom(2), not sys/random.h