Reorder personal data above passwords
It's an open question whether passwords belong under privacy settings at
all, but as long as we have them here, we shouldn't break the
association between ITP and the website data settings and the actual
website data by putting something unrelated in between them. Passwords
can go on the bottom.