build: Allow libportal support to be disabled
It isn't clear whether the API/ABI of libportal are entirely stable yet
(https://github.com/flatpak/libportal/issues/33) so it is not necessarily
appropriate for longer-term-supported OS distributions to include it.
When building a version of epiphany for a distribution package, which is
only intended to be packaged in a format other than as a Flatpak app,
libportal isn't necessary anyway.

libportal is also Linux-specific, so non-Linux OSs will likely want to
disable it (even if it might compile successfully).

Signed-off-by: Simon McVittie <smcv@debian.org>