gsb-service: don't crash when GSB database is broken
Currently there are four places where we assert the GSB database is not
broken. But some of these asserts are wrong. Two are clearly safe, and
one is clearly not (since it was actually hit). I think the other is
also not safe.

https://bugzilla.redhat.com/show_bug.cgi?id=1900656


(cherry picked from commit 3829d5687af829ef2e6ade8c77061ba052a2918a)