Make ephy-debug available without developer mode
There is zero value to gating these features behind developer mode. It's
just annoying to not have logging available outside developer mode. And
developer mode is incompatible with flatpak (because it forces the web
extension to be loaded from the source directory, but under flatpak it
needs to be loaded from the install directory), so unless we change
this, debugging flatpak builds is very hard.

Developer mode stays around to allow loading the web extension from the
source directory. It just won't gate debug features anymore.