search-engines-listbox: Move from in-code UI to UI files
Currently all this listbox and child search engine rows are built using
code packed in an enormous function. This makes it hard to make any
adjustment, and will make it harder to migrate to GTK4 later on.

So this commit separates the list box handling and UI from the rows one,
migrating both to UI files and enabling greater simplification of the
code.