prefs-dialog: Only save settings when closing the dialog
Currently, each time a modification is made the settings are written to
the disk.

But it isn't really needed to do it for each changes, saving only once
when closing the preferences is enough. This saves a lot of disk writes
when doing several changes for the same setting (like changing the name
of a search engine).