prefs: Fix setting the default search engine
Currently when setting an engine as the default one, it isn't always
kept once the preferences are reopened.

This is because when creating each row in the UI, the "active" property
of each radio button is default-initialized to NULL when constructing
it, but since we're already connected to the "clicked" signal it somehow
sends it for this FALSE state even though it didn't go to TRUE first.

So this commit makes sure that the clicked signal is a real click, by
checking for active state TRUE.

Fixes #1430