Fix regression preventing clicking links beneath the floating bar
When the floating bar covers a link at the bottom of the page, I should
be able to click the floating bar to click the link. This broke in !850.

There is some visual glitchiness where the bar briefly appears on the
left before being moved out of the way to the right, but I'm not sure
how to fix that