permissions-manager: stop using webkit_security_origin_is_opaque
I deprecated this function in WebKit r274270. It no longer does
anything, so let's stop using it. The original code here was not quite
right anyway, because it would return -1 whenever any origin is opaque.
This is not how comparison functions are supposed to work: it means a <
b and also b < a at the same time, which is outrageous. It just never
mattered because the security origins used here come from the main
resource URI, which is never opaque, so we can assert that there is
always a protocol and host. If we are looking up permissions for a
protocol that doesn't include a host component, like file:/// or
something, we have already lost.