Fix potential loss of session state when web process is unresponsive
If any web process is unresponsive when the window is closed using the
close button in the header bar -- not Ctrl+Q -- then the entire session
state is lost. Disaster! This happens to me three or four times per
year, and it is sad because it inevitably results in me forgetting
important things I had to do.

Anyway, the problem is our code to handle an unresponsive web process.
We'd like to cancel the JavaScript execution using
g_cancellable_cancel(), but it doesn't work because the underlying
WebKit operation is not actually cancellable, so instead we use a big
hammer and gtk_widget_destroy() the EphyWebView. Then when we wind up
closing the session later, there are no tabs open and nothing to save.
It was tricky to figure out how to fix this, because the only way to
truly cancel the operation is to destroy the web view, and if we do that
then we have to close the session first to avoid losing tabs, but the
user could legitimately choose to reject a modified forms close
confirmation warning, in which case they would wind up using Epiphany
after the session is closed, meaning all subsequent changes to open tabs
would be lost since they won't be saved after the session is closed.

Eventually I realized that we can just move the timeout from EphyWindow
-- where it's handled in two separate places, once for closing a single
tab, and once for closing an entire window -- down into EphyWebView
itself. This solves all our problems because now EphyWindow can assume
the JS execution always completes in a reasonable time.

I tested this by changing the call to
webkit_web_view_run_javascript_in_world() in
ephy_web_view_has_modified_forms() to execute a while(true) loop.
Without the other changes, the session state is lost if closing the
window using the close button in the header bar. With this patch, the
session state is no longer lost.

Fixes #1445