sync-dialog: say "GNOME Web" instead of "Web"
In the rest of our UI, we are careful to always brand the browser as
"Web." But that doesn't work very well here, since it's easy for users
to become confused as to the different between Web, Firefox, and the
web. Mozilla has very few requirements for use of the Firefox Sync
service other than to avoid confusing users, so let's err on the side of
caution and really spell out that we are GNOME Web here.

We might want to consider doing this elsewhere in the user interface as
well, but this particular use is more confusing than all others, so I
think this will suffice for now.