Fix crash when importing bookmarks from Firefox
The problem is the strings returned by get_firefox_profiles() are freed
with g_free(), which is correct, but we are actually returning pointers
into the middle of the allocated region, rather than pointers to the
start of the string. Truncating a string using pointer arithmetic is a
nice trick for unowned strings, but for owned strings it doesn't work.

https://bugzilla.redhat.com/show_bug.cgi?id=1946648


(cherry picked from commit 7e31b93d937625602e910d7397d00cc6082a37be)