suggestion-model: fix crash when typing in address bar
We have a refcounting error in ephy_suggestion_model_query_async(),
where we pass our task into query_collections_done() without reffing it.
Then the task is already destroyed when we go to form the history query,
and it has destroyed our QueryData with it, and so we have a UI process
crash when attempting to g_strsplit (data->query).

There's no need for each outstanding query to ref the task. Let's just
use one ref and drop it when the final query completes.

Fixes #1522

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/956>