Add unresponsive web process error page
WebKitGTK 2.33.1 allows us to detect and kill runaway web processes.
Let's take advantage of that.

The process is not detected as unresponsive for quite a while -- 10
seconds -- so there's no need for any secondary timeout at the Epiphany
level. Let's kill it immediately when WebKit says the web process has
been lost. If the WebKit timeout were shorter, then prompting the user
to decide whether to kill the page would make more sense, but 10 seconds
is long enough that any web content that hasn't returned to the main
loop deserves to be immediately killed. There's no excuse for that.