location-entry: Replace fake eventbox buttons with real ones
Currently the reader mode and bookmark icons in the location entry
are made clickable with the mouse using event boxes. So they are
kinda "fake buttons" which aren't focusable at all using the
keyboard.

So to make it easier to access them with the keyboard, and to have
to way of using their features (keyboard focus and keyboard shortcut),
this commit replaces the event boxes with proper buttons, styled in
a way where the buttons are fully invisible to make it look like you
click or focus a clickable icon, not a button.

This commit also adds a hover color for the entry icons.