bookmark-properties-grid: Make it properly focusable
Currently when using the Add/edit Bookmark popover or editing one
in the bookmarks list, pressing the Tab key doesn't focus anything
of the bookmarks properties, instead it closes the popover.

Looking at the UI file and source code, there's a whole lot of places
where the can-focus property is set, which in general isn't a good
sign because it's only needed as last resort. For example labels
had their can-focus to false, which is not needed unless the label
is made manually selectable. The close buttons of the tags (and the
tags themselves) both had it too set to False, which is strange
because they both are supposed to be clicked. This meant that while
they could be clicked with a mouse, they were never focusable with
the keyboard at all.

So this commit drops all the places where can-focus is set, which
makes pretty much everything work fine: all entries and buttons are
focusable and clickable with the keyboard.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/978>