bookmark-properties-grid: Change to a box container and rename widget
Currently the EphyBookmarkPropertiesGrid uses a GtkGrid to layout the
different widgets. But it only uses a single column, which means the
grid is basically useless as a GtkBox would do the job better without
requiring the different packing properties (as long as the childs are in
increasing order).

So this commit changes it to a GtkBox subclass and renames the widget
accordingly, without including "box" in the name as it's not useful.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/978>