clear-data-view: Drop useless flowbox
The flowbox was used when there was a second child, a comboboxtext, but it
is no longer useful because there's only a label as child now. It even made
the label look like it was focusable with the keyboard (with the focus outline),
as the GtkFlowBoxChild was focusable.

So this commit removes the useless flowbox.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/987>