resources/gtk: Remove plenty of useless (and harmless) can_focus
In the UI files there's a lot of places where can_focus is set, even
when it makes no sense to have it because by default it is set to False,
and is adapted by the widget implementations depending on whether it
makes sense to have it set. So for example tons of buttons or rows had
can-focus set to True, although it's already the default for those, or
set to False for labels. Some containers also had their can-focus set
to True/False although they almost never need focus (and have it set
to False by default too).

So this commit removes all those useless can-focus, but only the one
that shouldn't fix any keyboard focus issue.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/987>