embed: Only lowercase search URI's hostname when we know it's an URI
Currently if doing a search using a bang (e.g. "!ddg GNOME") the whole search
is lowercased. This is unexpected as one might want to include the correct
caps so that it lands on the right page straight away, without needing to
go through the search results page. This is the case for example if searching
"!wt Katze" (with the wiktionary), where it'll land on a page saying that
the "katze" page doesn't exist but instead there is the Katze page (in german
caps at the beginning of words is important for nouns), even though "Katze" was
already what was searched in Epiphany's search bar.

So this was a bug where it wanted to lowercase only the hostname part (like
GOOgle.com => google.com) but did it before it checked and processed the bang
search, which isn't a proper URI so it assumes the hostname is the whole string
and lowercases it.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/988>


(cherry picked from commit db602b13f35822eb409c3b32f8c8a7d59a7da251)