action-bar-end: Don't toggle bookmarks or downloads menu button if not shown
Currently if no downloads were made, using the Ctrl+Shift+Y shortcut shows
an unexpected empty popover. This does not happen in desktop mode for the
bookmarks popover because we don't hide it here.

So this commit makes those two buttons not get toggled when they are not actually
visible, to avoid seeing a popover coming out of nowhere (in mobile mode too).
This does not actually fix showing the popovers coming from the bottom bar in
mobile mode, but then it that you'd generally use a touchscreen without Ctrl
keys, so for the time being it's not a big deal though it could be fixed.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/990>