web-view: wait 10 seconds before killing unresponsive process
WebKit determines that the web process is hung when it has been more
than three seconds since it has responded to a ping from the UI process.
Turns out, this timeout is too low in practice because websites
routinely run JavaScript for longer than this. So let's wait 10 seconds
after the process is determined to be unresponsive, for 13 seconds
total, before we kill anything. My hope is that this makes the
unresponsive process killer less painful in practice.

We might still want to ask users before killing anything, #1561, but
this might suffice for now. Maybe.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/994>