about: Support dark style
Set color-scheme so we get light focus rects in dark variant.

Recikir icons, since they are monochrome here (well, almost but it's close
enough) and we don't know beforehand what the icons will be the app icon
in particular), be careful and invert them as:
  - set their brightness to 0, making sure their color is black
  - then use invert with a carefully chosen value

Don't set titles to the wrong color - text-color is for text areas, this
one is fg-color.