error: Make error pages work in dark
Same as the last commit, and a few more adjustments.

Recoloring icons is easy when they are monochrome. Unfortunately, here we
have red icons. So here we have to use complex and fragile filters, assume
the source icon is #cc0000 and carefully manipulate it to get the precise
color we want.

We also have to move the style class for it to work. While it could be both
on the icon and the title, it's easier to add it to the parent instead.

Fix the details expander.