Stop using about:blank for new tabs
about:blank is still white even when the app is using dark style.

Go with the same approach as Safari and introduce a new page for new tab
instead: about:newtab. As a bonus, we can give it a proper title instead
of "blank page".

Fixes https://gitlab.gnome.org/GNOME/epiphany/-/issues/1555