embed: Inline EphyEmbedEvent
It was probably more useful when it was written, but by now it's a thin
wrapper around WebKitHitTestResult that proxies everything to it and
doesn't do much else. Also contains a GdkEvent and proxies its fields
but it's unused.

So just inline it instead.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1034>