encoding-dialog: Port to GtkBox
There was a period where GtkGrid looked like it's going to be the future,
along with a misleading mention that boxes are going to be removed. In
practice, that was about HBox and VBox, not Box. Meanwhile, this layout has
no reason whatsoever to use grid and it just complicates it.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1038>