Detect Snap sandbox in addition to Flatpak
Snaps use the same portals as Flatpak, so we should detect whether we're
running as a Snap when deciding whether to use portals or disable
features that are impossible in these different but similar sandboxing
technologies.

This is mostly academic because the Snap of Epiphany appears to be
outdated, but it is more correct so let's do it.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1041>