pdf-handler: properly encode filename before inserting to HTML
The file's name is suggested by the server, and could be malicious. We
don't want it to be able to escape the HTML attribute context.

The file data should already be safe because it is base-64 encoded. Here
I'm just adjusting the code style to match what I've done for the
filename.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1045>