view-source-handler: encode data passed to highlight.js
The actual data here should be good already because it gets escaped by
GLib, but this function is really designed for use in XML, so let's
switch to the simpler Epiphany function designed for anti-XSS to make it
more clear what's going on here.

The URL is probably vulnerable, though, since a malicious URL could
conceivably try to escape the HTML entity context. Encode that.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1045>