ci: don't check for memory leaks, use-after-free, double free
This is unfortunate, but the checker is too dumb. It also has no way to
suppress false positives, so we either have to make undesirable changes
to the code, or else disable the checker entirely.

We only really want to disable the memory leak checker, but there's no
way to do that without also disabling the check for use-after-free or
double free, so here we are.

My opinion of scan-build has declined drastically today. This is sad.
But it seems like the best we can do.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1045>