Fix reload buttons on error pages
The encoded URL here does not work. And we cannot reload via the web
process, because the window.location is about:blank for alternate HTML,
so we'll have to send a message to the UI process to have it do so
instead.

Fixes #1663

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1050>