about-handler: don't encode app ID (or install date)
We are overencoding here. Epiphany is not prepared to handle the encoded
app ID, and it is not necessary to encode in the first place because the
app ID is trusted to be a valid GApplication ID, which cannot contain
nasty characters.

However, encoding the URLs here really is necessary, because they really
could contain nasty content.

Fixes #1665

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1050>