prefs/search-engine-row: Move a comment back to its correct location
This comment was there because previously, when the search engine row
was built in a 100s lines function directly in ephy-search-engine-list-box.c,
it was malloc'ed at the beginning of the function and memset'ed to 0
at the beginning of the loop. But since that didn't make sense at all
retrospectively when moving the row to a separate widget and file,
I just moved it to a local variable in the loop, and without malloc. It
seems like I didn't at that time move the comment too, so this commit
fixes it.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1052>