prefs/search-engine-row: Use the "fast" variant of g_utf8_to_ucs4()
We already hard error out anyway when there's an invalid UTF-8 character,
and the documentation of g_utf8_to_ucs4_fast() states that this function
is twice as fast as the non "fast" one. So better use it here then.
Also, since we know at the end that we're getting valid UTF-8, let's assert
instead of g_error() with a not very useful error message.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1052>