profile-migrator: Allow specifying --do-step-n <n> when n is the same as the last migration version
Currently it's impossible to manually run the profile migrator to debug
it with gdb using the --do-step-n <n> argument where n is the latest
migration version, as the migrator refuses to run the migration when
the latest migration version matches do_step_n. I don't see how that
makes sense, so this commit drops this restriction.

This shouldn't affect the normal course of migration as done by the
main epiphany binary, as it already checks if the asked migration
version matches the one in the .migrated file, in the
ephy_profile_utils_do_migration () function.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1052>