Fix web app GApplication ID calculation
In get_gapplication_id_from_id() we sometimes make the GApp ID
"org.gnome.Epiphany.WebApp-<checksum>". This is then used as the profile
directory name. The profile directory is then used to calculate the id
in ephy_web_application_setup_from_profile_directory(). But in
get_app_id_from_gapplication_id() we return whatever's after the
"WebApp-" prefix, since there's no way to recover the "normalized name"
part of the ID. So, the functions that handle the ID need to work when
it's just the checksum part without the normalized name.

This is all very confusing and we should probably find a way to make it
all simpler some day :/

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1063>