Migrate web apps to portal locations
This is yet another migration of web apps. This time we are changing the
GApplication ID to make it simpler and more D-Bus friendly. We also have
to ensure the GApp ID has the app ID as a prefix to be able to use the
dynamic launcher portal (and in order to own the right bus name when the
web app is run). We also migrate the web apps' desktop files and
icons to the locations they'd be in if we had used the portal to create
them, so that going forward they will all be in one place and things
should be simpler.