Use portals for web app install/uninstall
This makes use of the new dynamic launcher portal
(https://github.com/flatpak/xdg-desktop-portal/pull/696) to install and
uninstall web apps. This means we are able to support web apps now in
Flatpak'd Epiphany. Unfortunately it's hard to test though so for
now remove the unit tests.

Disable the app manager feature under Flatpak for now; a later commit
will re-enable it.