search-engines: Port whole EphySearchEngineManager code to GListModel
This commit ports EphySearchEngineManager to being a proper GListModel,
with a separate EphySearchEngine object for each search engines. That makes
the code overall way cleaner and easier to work with, as there's no need
to keep track of the old_name and saved_name or whatever horrible thing
that I needed to do, since now we just need to keep around the EphySearchEngine
object we're displaying in this preferences row, and do changes on it directly.
That did require quite a few changes all around the code base to adapt to all
API changes, but it is definitely worth it.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1055>