general-page: Stop using GtkFileChooserButton
Unfortunately, we have to reimplement fetching the display name. GTK does
a quite elborate thing here, we don't need to be that fancy, but would be
nice to have at least Downloads and Home be named as such and not as your
username. Additionally, since we handle Desktop, let's special case that as
well.

Make this all a helper function, as we'll need to do it in the download
dialog as well.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1072>