synced-tab-dialog: Fix named icons
The computer icon was still dark in dark variant. The placeholder icon was
just missing because we've moved the icon in resources at some point. Since
that's not what we use for tabs anyway, fix that as well.

While we're here, fix a leak - we were passing
ephy_pixbuf_get_from_surface_scaled() results into the tree store and
forgetting about them.