Fix web app icon url vs path confusion
Confusingly, the icon_url member of EphyWebApplication is actually a
path not a URL or URI. The only place where it gets set to a URI is in
prefs-general-page.c when the user changes the icon, and there it is
happening erroneously since all the other code that deals with it
assumes it is a path. So, rename the struct member and ensure all the
use of it treats it as a path.