Simplify bookmark opening
Remove the win.bookmark-open action as it's only used from one place.
Instead, just inline that code.

Since this is the last use of ephy_link_flags_from_current_event(), inline
it as well. While it means we're still using ephy_gui_get_current_event(),
at this point it's very close to the actual event handling and can be
replaced relatively easily in GTK4 - unlike when it's in a GAction.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1081>