Disable all Safe Browsing code at build time
Currently we have a runtime check to disable creation of the
EphyGSBService if the API key is not provided at build time. I think
that is sufficient, but since this code should be unusable if API key is
not provided, let's not build it at all for some extra assurance that it
won't run.

Related: #1721