Switch Tech Preview to use GTK 4 branch config/fedora: drop f33 since its EOL
Signed-off-by: Jordan Petridis <jordan@centricular.com>

This will allow some extra testing for the GTK 4 branch before we pull
the trigger to merge it into master. It's a lot easier to revert this
commit than it is to revert such a merge.

Note this also temporarily removes elementary OS support. It will be
welcome back once it's ready for GTK 4.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1100>