Canary: Enable the new libva GStreamer decoders
The WebKit SDK now ships the gstva plugin, which aims to be a replacement for
gstreamer-vaapi. The elements are currently down-ranked but should be usable,
testing was performed on a machine with a Radeon rx 560 GPU, but should also
work on Intel GPUs (YMMV).

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1105>