file-helpers: always delete tmp directory
The KEEP_DIRECTORY flag is being misinterpreted as indication that
the profile directory should not be deleted. In fact, it controls
whether the Epiphany temporary directory is deleted. And that should
always be deleted, because leaving temp directories around is
unfriendly, so get rid of it altogether.

Fixes #1768

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1108>