password-manager: Fix crash when deleting all passwords
This broke in 7d080a31. The callback here is not prepared for there to
be no async data, but the async data is only required when clearing a
single password, not when clearing all of them. The GTask is optional
even if there is async data. We're just going to need a bunch of
conditionals here.

Fixes #1773

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1109>


(cherry picked from commit 1e2a52481b7755fd6fe373efc8cd325f6c4e8960)