user-agent: restore application branding
This is part of a deal with Google to stop accounts.google.com from
discriminating against us. It's a very risky change, but we'll try it
for now and make a list of websites that it breaks.

Historical background, in order of recency:

https://bugs.webkit.org/show_bug.cgi?id=240041
https://bugs.webkit.org/show_bug.cgi?id=222978
https://bugs.webkit.org/show_bug.cgi?id=222039#c5
https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/595

Also, remove the web app branding, since the risk outweighs the reward
and I don't want a separate set of bugs in web app mode. I believe only
one website was using this:

https://bugzilla.gnome.org/show_bug.cgi?id=767123

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1121>