Revert "user-agent: restore application branding"
This reverts commit 3e59a634a87baa8f8a77d63f4a4009433a96cbcb.

It breaks id.spectrum.net, which shows an unsupported browser warning,
even if I modify WebKit to place the application branding in the middle
of the header instead of at the end. This is just the first such website
of many. We're not going to be able to set application branding in
general.

If particular websites require a branding quirk, we'll have to do that
in WebKit.