Fix crashes and misbehavior when importing bookmarks
We are trying to use the index of the selected item to decide what
action to perform, but this doesn't work because not all items are
actually present in the combo box. E.g. if there is no Firefox profile,
it won't be present in the list. This causes the wrong codepath to be
selected. In this bug report, the user complains that it crashes due to
not finding any Firefox profiles when importing bookmarks from Chrome.

Fixes #1800


(cherry picked from commit a22f87c895ee36b7a7905a91ac181d0a6522170f)