downloads-popover: fix hang when using Clear All button
If the Clear All button is pressed while a download is still active, we
loop forever because the loop's control condition just removes the first
item from the list and continues until all have been removed, but only
downloads that are not active get removed. We need to track how many
active downloads we have skipped to ensure we terminate.

Fixes #1803

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1131>