WebExtensions: Unify web-process-extensions
Previously there were two web process extensions that tried to
not overlap with eachother however this design doesn't make sense.

Any tab can load web-extension pages. The only distinction is the URI they load.

So this combines them and depending on the URI loaded it will install the correct APIs.