window-commands: work around GTK native dialog hang
When importing bookmarks, we have:

 (1) EphyWindow, transient parent of
 (2) GtkDialog to select where to import bookmarks from, transient
parent of
 (3) The native file chooser dialog (in case of GVDB or HTML import)

Sadly, opening a native dialog on a modal dialog hangs the application,
gtk#5011. To dodge this problem, we can close the second dialog
dialog when opening the third dialog, and make the EphyWindow the
transient parent of the third dialog. This is inconvenient because it
means you're all the way back to step one if you decide to cancel the
third dialog and want to go back to the second dialog, but it's better
than a hang.

Exporting bookmarks does not have this problem, because there we skip
step 2.

Importing passwords does not have this problem, because there we don't
have a native file chooser. (At least, I think so. Didn't test this
because I do not have Chrome installed.)