Fix loss of tab history if placeholder page is not removed
When starting Epiphany, we load an empty HTML document in all tabs
except the currently-presented tab in order to start quickly and avoid
excess memory usage. The real document is loaded later when the user
clicks on the tab for the first time after starting the browser. But
there's a problem with this strategy: if the user never clicks on the
tab, then the placeholder document is still there when we quit, and we
accidentally save the *placeholder* into the session state instead of
the real state of the tab that was restored from the previous session.
The most obvious impact of this is you lose the contents of the
back/forward list.

This fixes the issue by ensuring we check to see if delayed session
state is available when saving the session. If so, the web view is
displaying a placeholder document and we want to save the delayed state,
not the placeholder state.

Fixes #1622

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1191>